# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Prisma Embedded Codes — Authentication & Authorization System API
  description: Security-hardened REST API for user registration, authentication, MFA (TOTP/Email), token rotation, and RBAC authorization.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development Server
paths:
...
  /auth/refresh:
    post:
      summary: Rotate refresh token
      description: Validates current refresh token and returns a new rotated token pair. Protects against reuse theft.
      security:
        - CookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Optional body fallback if cookies are blocked
      responses:
        '200':
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=new-opaque-token; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/401Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
    ErrorResponse:
      type: object
      required:
        - statusCode
        - error
        - code
        - message
        - requestId
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid inputs provided
        details:
          type: object
        requestId:
          type: string
          format: uuid
  responses:
...
    401Unauthorized:
      description: Authentication tokens are missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...